requirements.txt
agent-platform-sdk @ git+https://github.com/GenAISuperman/agent-platform-sdk.git@main
langgraph>=0.2
langchain>=0.3
langchain-openai>=0.2
fastapi>=0.115
uvicorn>=0.32
opentelemetry-api>=1.25
opentelemetry-sdk>=1.25
opentelemetry-exporter-otlp>=1.25
azure-identity>=1.17
azure-keyvault-secrets>=4.8
pydantic>=2.0
python-dotenv>=1.0
